feat(packaging): Add modern packaging and restructure layout
Introduce a pyproject.toml to make the project installable via
pip and to define a console script entry point.

Restructure the source files from a flat layout into a proper
package directory (`picasa2digikam/`) to enable a clean build
with setuptools. This allows the tool to be installed and run
directly from the command line.
