+++ .gitignore
@@ -8,3 +8,4 @@ tmp
 *.py[cod]
 
 test.sh
+picasa2digikam*-info/
