+++ picasa2digikam/digikam_db.py
