+++ picasa2digikam/rect64.py
